+++ core/src/tech/team28/heslingtonhustle/GameManager.java
@@ -3,10 +3,10 @@ package tech.team28.heslingtonhustle;
 import com.badlogic.gdx.utils.Array;
 
 public class GameManager {
-    static final int GAME_WIDTH = 1000;
-    static final int GAME_HEIGHT = 1000;
+    static final float GAME_WIDTH = 1000f;
+    static final float GAME_HEIGHT = 1000f;
     // View width determined by window aspect ratio
-    static final int VIEW_HEIGHT = 1000;
+    static final float VIEW_HEIGHT = 1000f;
 
     public void setPlayer(Player player) {
         this.player = player;
